Split out wait-normal-finish tests.
