+++ computation-slurm-test/src/test/java/com/powsybl/computation/slurm/SlurmUnitTests.java
@@ -9,9 +9,6 @@ package com.powsybl.computation.slurm;
 import com.powsybl.commons.config.ModuleConfig;
 import com.powsybl.commons.config.YamlModuleConfigRepository;
 import com.powsybl.computation.*;
-import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
-import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
-import org.apache.commons.io.IOUtils;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Ignore;
@@ -20,22 +17,18 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import java.util.stream.IntStream;
-import java.util.zip.GZIPOutputStream;
 
+import static com.powsybl.computation.slurm.CommandExecutionsTestFactory.*;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
@@ -45,18 +38,20 @@ import static org.junit.Assert.assertTrue;
 @Ignore
 public class SlurmUnitTests {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(SlurmUnitTests.class);
-    private static final ExecutionEnvironment EMPTY_ENV = new ExecutionEnvironment(Collections.emptyMap(), "unit_test_", false);
+    static final Logger LOGGER = LoggerFactory.getLogger(SlurmUnitTests.class);
+    static final ExecutionEnvironment EMPTY_ENV = new ExecutionEnvironment(Collections.emptyMap(), "unit_test_", false);
 
-    private ModuleConfig config;
+    ModuleConfig config;
+    SlurmComputationConfig slurmConfig;
 
-    private volatile boolean failed = false;
+    volatile boolean failed = false;
 
     @Before
     public void setup() {
         YamlModuleConfigRepository configRepository = new YamlModuleConfigRepository(Paths.get("src/test/resources/config.yml"));
         config = configRepository.getModuleConfig("slurm-computation-manager")
                 .orElseThrow(() -> new RuntimeException("Config.yaml is not good. Please recheck the config.yaml.example"));
+        slurmConfig = generateSlurmConfigWithShortScontrolTime(config);
         // TODO prepare myapps if necessary
     }
 
@@ -64,7 +59,7 @@ public class SlurmUnitTests {
         return new SlurmComputationConfig.SshConfig(config.getStringProperty("hostname"), 22, config.getStringProperty("username"), config.getStringProperty("password"), 10, 5);
     }
 
-    private static SlurmComputationConfig generateSlurmConfig(ModuleConfig config) {
+    static SlurmComputationConfig generateSlurmConfig(ModuleConfig config) {
         return new SlurmComputationConfig(generateSsh(config), config.getStringProperty("remote-dir"),
                 Paths.get(config.getStringProperty("local-dir")), 5, 4);
     }
@@ -72,7 +67,7 @@ public class SlurmUnitTests {
     /**
      * Returns a config in which scontrol monitor runs every minute
      */
-    private static SlurmComputationConfig generateSlurmConfigWithShortScontrolTime(ModuleConfig config) {
+    static SlurmComputationConfig generateSlurmConfigWithShortScontrolTime(ModuleConfig config) {
         return new SlurmComputationConfig(generateSsh(config), config.getStringProperty("remote-dir"),
                 Paths.get(config.getStringProperty("local-dir")), 5, 1);
     }
@@ -120,111 +115,6 @@ public class SlurmUnitTests {
         }
         // assert on main thread
         assertFalse(failed);
-    }
-
-    @Test
-    public void testSimpleCmdWithCount() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "simpleCmdWithCount");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-            @Override
-            public List<CommandExecution> before(Path path) {
-                return simpleCmdWithCount(7);
-            }
-
-            @Override
-            public Void after(Path workingDir, ExecutionReport report) {
-                try {
-                    return super.after(workingDir, report);
-                } catch (IOException e) {
-                    failed = true;
-                }
-                return null;
-            }
-
-            private List<CommandExecution> simpleCmdWithCount(int count) {
-                return CommandExecutionsTestFactory.simpleCmdWithCount(count);
-            }
-        };
-        baseTest(testAttribute, supplier);
-    }
-
-    @Test
-    public void testLongTask() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "longTask");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-            @Override
-            public List<CommandExecution> before(Path workingDir) {
-                return longProgram(10);
-            }
-        };
-        baseTest(testAttribute, supplier);
-    }
-
-    @Test
-    public void testMyEchoSimpleCmd() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "myEchoSimpleCmdWithUnzipZip");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-            @Override
-            public List<CommandExecution> before(Path workingDir) {
-                generateZipFileOnRemote("in0", workingDir.resolve("in0.zip"));
-                generateZipFileOnRemote("in1", workingDir.resolve("in1.zip"));
-                generateZipFileOnRemote("in2", workingDir.resolve("in2.zip"));
-                return CommandExecutionsTestFactory.myEchoSimpleCmdWithUnzipZip(3);
-            }
-
-            @Override
-            public Void after(Path workingDir, ExecutionReport report) {
-                try {
-                    super.after(workingDir, report);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                    failed = true;
-                }
-                Path out2 = workingDir.resolve("out2.gz");
-                System.out.println("---------" + testAttribute.testName + "----------");
-                System.out.println("out2.gz should exists:" + Files.exists(out2));
-                failed = !Files.exists(out2);
-                System.out.println("------------------------------------");
-                return null;
-            }
-        };
-        baseTest(testAttribute, supplier);
-    }
-
-    private static void generateZipFileOnRemote(String name, Path dest) {
-        try (InputStream inputStream = SlurmUnitTests.class.getResourceAsStream("/afile.txt");
-             ZipArchiveOutputStream zos = new ZipArchiveOutputStream(Files.newOutputStream(dest))) {
-            ZipArchiveEntry entry = new ZipArchiveEntry(name);
-            zos.putArchiveEntry(entry);
-            IOUtils.copy(inputStream, zos);
-            zos.closeArchiveEntry();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Test
-    public void testGroupCmd() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "groupCmd");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-            @Override
-            public List<CommandExecution> before(Path workingDir) {
-                return CommandExecutionsTestFactory.groupCmd();
-            }
-        };
-        baseTest(testAttribute, supplier);
-    }
-
-    @Test
-    public void testTwoSimpleCmd() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "twoSimpleCmd");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-            @Override
-            public List<CommandExecution> before(Path workingDir) {
-                return twoSimpleCmd();
-            }
-        };
-        baseTest(testAttribute, supplier);
     }
 
     @Test
@@ -330,98 +220,6 @@ public class SlurmUnitTests {
     }
 
     @Test
-    public void testFilesReadBytes() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "filesReadBytes");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-
-            static final int COUNT = 10;
-
-            @Override
-            public List<CommandExecution> before(Path workingDir) {
-                Command command = new SimpleCommandBuilder()
-                        .id("echo")
-                        .program("echo asdf >")
-                        .arg(i -> i + ".txt")
-                        .build();
-                return Collections.singletonList(new CommandExecution(command, COUNT));
-            }
-
-            @Override
-            public Void after(Path workingDir, ExecutionReport report) throws IOException {
-                IntStream.range(0, COUNT).forEach(i -> {
-                    try {
-                        System.out.println("i : " + i);
-                        byte[] bytes = Files.readAllBytes(workingDir.resolve(i + ".txt"));
-                        byte[] bytes1 = Files.readAllBytes(workingDir.resolve("echo_" + i + ".out"));
-                        byte[] bytes2 = Files.readAllBytes(workingDir.resolve("echo_" + i + ".err"));
-                        System.out.println(bytes.length);
-                        System.out.println(bytes1.length);
-                        System.out.println(bytes2.length);
-                    } catch (IOException e) {
-                        failed = true;
-                    }
-                });
-                return super.after(workingDir, report);
-            }
-        };
-        baseTest(testAttribute, supplier);
-    }
-
-    @Test
-    public void testMd5sumLargeFile() throws InterruptedException {
-        TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "md5sumLargeFile");
-        Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
-            @Override
-            public List<CommandExecution> before(Path workingDir) {
-                long startDump = System.nanoTime();
-                generateGzFileOnRemote(2, workingDir.resolve("2GFile.gz"));
-                generateGzFileOnRemote(4, workingDir.resolve("4GFile.gz"));
-                LOGGER.info("Dump two files in {} ms", TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startDump));
-                return md5sumLargeFile();
-            }
-
-            @Override
-            public Void after(Path workingDir, ExecutionReport report) throws IOException {
-                List<String> actual2GMd5 = Files.readAllLines(workingDir.resolve("c1_0.out"));
-                List<String> actual4GMd5 = Files.readAllLines(workingDir.resolve("c2_0.out"));
-                String expected2GMd5 = "1ea9851f9b83e9bd50b8d7577b23e14b  2GFile";
-                String expected4GMd5 = "bbe2b516d690f337d8f48fc03db99c9a  4GFile";
-                if (!Objects.equals(actual2GMd5.get(0), expected2GMd5)) {
-                    failed = true;
-                    System.out.println("  actual:" + actual2GMd5.get(0));
-                    System.out.println("expected:" + expected2GMd5);
-                }
-                if (!Objects.equals(actual4GMd5.get(0), expected4GMd5)) {
-                    failed = true;
-                    System.out.println("  actual:" + actual4GMd5.get(0));
-                    System.out.println("expected:" + expected4GMd5);
-                }
-                return null;
-            }
-        };
-        baseTest(testAttribute, supplier);
-    }
-
-    private void generateGzFileOnRemote(int sizeGb, Path dest) {
-        StringBuilder sb = new StringBuilder();
-        IntStream.range(0, 1024).forEach(i ->
-                IntStream.range(0, 128).forEach(j -> sb.append("KKKKKKKK"))); // 1mb
-        String oneMb = sb.toString();
-        try (OutputStream outStream = new GZIPOutputStream(Files.newOutputStream(dest))) {
-            IntStream.range(0, 1024 * sizeGb).forEach(i -> {
-                try {
-                    outStream.write(oneMb.getBytes());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                    failed = true;
-                }
-            });
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Test
     public void makeSlurmBusy() throws InterruptedException {
         TestAttribute testAttribute = new TestAttribute(Type.TO_WAIT, "deadline");
         Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
@@ -441,7 +239,7 @@ public class SlurmUnitTests {
         Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
             @Override
             public List<CommandExecution> before(Path workingDir) {
-                return CommandExecutionsTestFactory.longProgram(10);
+                return longProgram(10);
             }
         };
         ComputationParametersBuilder builder = new ComputationParametersBuilder();
@@ -471,7 +269,7 @@ public class SlurmUnitTests {
         Supplier<AbstractExecutionHandler<Void>> supplier = () -> new AbstractExecutionHandler<Void>() {
             @Override
             public List<CommandExecution> before(Path workingDir) {
-                return CommandExecutionsTestFactory.longProgram(10);
+                return longProgram(10);
             }
         };
         ComputationParameters parameters = new ComputationParametersBuilder().setTimeout("longProgram", 60).build();
@@ -518,20 +316,6 @@ public class SlurmUnitTests {
         }
     }
 
-    private static List<CommandExecution> twoSimpleCmd() {
-        Command command1 = new SimpleCommandBuilder()
-                .id("simpleCmdId")
-                .program("sleep")
-                .args("10s")
-                .build();
-        Command command2 = new SimpleCommandBuilder()
-                .id("cmd2")
-                .program("echo")
-                .args("hello", ">", "output")
-                .build();
-        return Arrays.asList(new CommandExecution(command1, 1), new CommandExecution(command2, 1));
-    }
-
     private static List<CommandExecution> invalidProgram() {
         Command cmd = new SimpleCommandBuilder()
                 .id("invalidProgram")
@@ -558,15 +342,6 @@ public class SlurmUnitTests {
 
     private static List<CommandExecution> invalidProgramInList() {
         return Arrays.asList(invalidProgram().get(0), CommandExecutionsTestFactory.simpleCmd().get(0));
-    }
-
-    private static List<CommandExecution> longProgram(int seconds) {
-        Command command = new SimpleCommandBuilder()
-                .id("longProgram")
-                .program("sleep")
-                .args(seconds + "s")
-                .build();
-        return Collections.singletonList(new CommandExecution(command, 1));
     }
 
     private static List<CommandExecution> longProgramToCancel() {
@@ -613,23 +388,6 @@ public class SlurmUnitTests {
         return Arrays.asList(new CommandExecution(command1, 1),
                 new CommandExecution(command2, 5),
                 new CommandExecution(command3, 2));
-    }
-
-    private static List<CommandExecution> md5sumLargeFile() {
-        Command command1 = new SimpleCommandBuilder()
-                .id("c1")
-                .program("md5sum")
-                .inputFiles(new InputFile("2GFile.gz", FilePreProcessor.FILE_GUNZIP))
-                .args("2GFile")
-                .build();
-        Command command2 = new SimpleCommandBuilder()
-                .id("c2")
-                .inputFiles(new InputFile("4GFile.gz", FilePreProcessor.FILE_GUNZIP))
-                .program("md5sum")
-                .args("4GFile")
-                .build();
-        return Arrays.asList(new CommandExecution(command1, 3),
-                new CommandExecution(command2, 1));
     }
 
     private enum Type {
